+++ src/frontend/apps/impress/src/components/quick-search/QuickSearchInput.tsx
@@ -1,6 +1,5 @@
-import { Loader } from '@gouvfr-lasuite/cunningham-react';
 import { Command } from 'cmdk';
-import { ReactNode } from 'react';
+import { PropsWithChildren } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { HorizontalSeparator } from '@/components';
@@ -9,19 +8,16 @@ import { useCunninghamTheme } from '@/cunningham';
 import { Box } from '../Box';
 import { Icon } from '../Icon';
 
-type Props = {
-  loading?: boolean;
+type QuickSearchInputProps = {
   inputValue?: string;
   onFilter?: (str: string) => void;
   placeholder?: string;
-  children?: ReactNode;
   withSeparator?: boolean;
   listId?: string;
   onUserInteract?: () => void;
   isExpanded?: boolean;
 };
 export const QuickSearchInput = ({
-  loading,
   inputValue,
   onFilter,
   placeholder,
@@ -30,7 +26,7 @@ export const QuickSearchInput = ({
   listId,
   onUserInteract,
   isExpanded,
-}: Props) => {
+}: PropsWithChildren<QuickSearchInputProps>) => {
   const { t } = useTranslation();
   const { spacingsTokens } = useCunninghamTheme();
 
@@ -52,14 +48,7 @@ export const QuickSearchInput = ({
         $gap={spacingsTokens['2xs']}
         $padding={{ horizontal: 'base', vertical: 'sm' }}
       >
-        {!loading && (
-          <Icon iconName="search" $variation="secondary" aria-hidden="true" />
-        )}
-        {loading && (
-          <div>
-            <Loader size="small" />
-          </div>
-        )}
+        <Icon iconName="search" $variation="secondary" aria-hidden="true" />
         <Command.Input
           autoFocus={true}
           aria-label={t('Quick search input')}