+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/custom-inline-content/Interlinking/SearchPage.tsx
@@ -181,6 +181,7 @@ export const SearchPage = ({
             setSearch(value);
           }}
           onKeyDown={handleKeyDown}
+          autoComplete="off"
         />
       </Box>
       <Box