+++ src/frontend/apps/impress/src/features/docs/doc-search/components/index.ts
@@ -1,3 +1,4 @@
+export * from './DocSearchContent';
 export * from './DocSearchModal';
 export * from './DocSearchFilters';
 export * from './DocSearchSubPageContent';